+++ preload.js
@@ -5,8 +5,6 @@
 //https://www.electronjs.org/docs/latest/tutorial/tutorial-preload
 // const { contextBridge } = require('electron');
 
-
-
 // contextBridge.exposeInMainWorld('Sentry', {
 //     captureException: (exception, context = undefined) => Sentry.captureException(exception, context),
 //     captureMessage: (message) => Sentry.captureMessage(message)